requirements.txt
finviz
yfinance
pandas
pandas-ta
alpaca-py
python-dotenv